# Streamline HTTP header/trailer validation and conversion