Motivation:

Currently, user-provided HTTP headers and trailers are consumer by
`HttpRequestSubscriber` and `HttpResponseSubscriber`. The following is
the basic header/trailer validation and conversion process:

1. `Http*Subscriber` merges the `Request/ResponseHeaders` and
   `ctx.additional*Headers()` while excluding some prohibited headers.
2. `Http*Subscriber` passs the merged headers to `HttpObjectEncoder`.
3. `HttpObjectEncoder` converts the merged headers while excluding some
   more prohibited headers, using `ArmeriaHttpUtil`.

(Pretty much same for trailers)

Instead, we could streamline and optimize this process like the following:

1. `Http*Subscriber` passes `Request/ResponseHeaders` and
   `ctx.additional*Headers()`, without any validation or exclusion.
2. `HttpObjectEncoder` converts the given headers while excluding
   prohibited headers and merging `Request/ResponseHeaders` and
   `ctx.additional*Headers()`.

This way, we could reduce the number of unnecessary instantiation of
immutable headers and their builders.

Modifications:

- Move and merge response header conversion process from
  `HttpResponseSubscriber` to `ArmeriaHttpUtil.toNettyHttp1/2()`
- Move and merge request header conversion process from
  `HttpRequestSubscriber` to `ArmeriaHttpUtil.toNettyHttp1/2()`
- Add parameters to `HttpObjectEncoder.writeHeaders()`:
  - `HttpHeaders additionalHeaders`
  - `HttpHeaders additionalTrailers`
- Split `Http1ObjectEncoder` into:
  - `ServerHttp1ObjectEncoder`
  - `ClientHttp1ObjectEncoder`
- Split `Http2ObjectEncoder` into:
  - `ServerHttp2ObjectEncoder`
  - `ClientHttp2ObjectEncoder`
- Split `ArmeriaHttpUtil.toNettyHttp1()` into:
  - `toNettyHttp1ServerHeader()`
  - `toNettyHttp1ServerTrailer()`
  - `toNettyHttp1ClientHeader()`
  - `toNettyHttp1ClientTrailer()`
- Split `ArmeriaHttpUtil.toNettyHttp2()` into:
  - `toNettyHttp2ServerHeader()`
  - `toNettyHttp2ServerTrailer()`
  - `toNettyHttp2ClientHeader()`
  - `toNettyHttp2ClientTrailer()`

Result:

- Better HTTP header validation/conversion performance
- Fixes #2096