@jyblue 2020-01-05
* master branch result
```
Benchmark                                                              Mode  Cnt      Score     Error  Units
HttpResponseHeaderConversionBenchmark.http1HeaderConversionBenchmark  thrpt    5  13332.762 ± 116.813  ops/s
HttpResponseHeaderConversionBenchmark.http2HeaderConversionBenchmark  thrpt    5  14736.343 ± 270.169  ops/s
HttpRequestHeaderConversionBenchmark.http1HeaderConversionBenchmark   thrpt    5  15287.277 ± 436.491  ops/s
HttpRequestHeaderConversionBenchmark.http2HeaderConversionBenchmark   thrpt    5  14750.212 ± 172.706  ops/s
```
* `c70af23` branch result
```
Benchmark                                                              Mode  Cnt      Score     Error  Units
HttpResponseHeaderConversionBenchmark.http1HeaderConversionBenchmark  thrpt    5  15087.681 ± 192.233  ops/s
HttpResponseHeaderConversionBenchmark.http2HeaderConversionBenchmark  thrpt    5  14797.992 ± 156.278  ops/s
HttpRequestHeaderConversionBenchmark.http1HeaderConversionBenchmark   thrpt    5  16260.746 ± 265.471  ops/s
HttpRequestHeaderConversionBenchmark.http2HeaderConversionBenchmark   thrpt    5  14973.532 ± 317.015  ops/s
```