@trustin 2020-01-29
We could further split more methods like `HttpObjectEncoder.writeHeaders()`, but I guess it's not part of this PR because validation and conversion is now single pass already.