@jyblue 2020-01-29
> Had a chat with @trustin, and we thought that `Http1ObjectEncoder` and `Http2ObjectEncoder` have so many branches right now so it's really hard to review and error-prone.
> In order to simplify that, we can perhaps split the encoders to `ClientHttp1ObjectEncoder`, `ServerHttp1ObjectEncoder`, `ClientHttp2ObjectEncoder` and `ServerHttp1ObjectEncoder`. Then, we might split the `ArmeriaHttpUtil` as well.
> Of course, I'm not sure this is the best we can do but I think it's worth to try it. Also, this PR does not bring any breaking changes to us, I think we can just take our time to improve this and do the best we can do.
> WDYT?

I agree. Once we split `Http1ObjectEncoder`, `Http2ObjectEncoder` and `ArmeriaHttpUtil`, we might remove many `if`s and codes will be more readable than now I think :thinking: