@jyblue 2020-01-29
Thank you reviewers, here are some updates.
I will commit more updates with remaining items fixed.