@trustin 2020-01-31
Would you mind posting a new perf test result after the changes so far?