@jyblue 2020-01-31
Thank you for quick fix. I will post a new bench result soon.