@jyblue 2020-01-31
This is new performance test result.

master `90b08a8`
```
Benchmark                                                             Mode  Cnt      Score     Error  Units
HttpRequestHeaderConversionBenchmark.http1HeaderConversionBenchmark   thrpt    5  14715.288 ± 318.914  ops/s
HttpRequestHeaderConversionBenchmark.http2HeaderConversionBenchmark   thrpt    5  14561.032 ± 285.466  ops/s
HttpResponseHeaderConversionBenchmark.http1HeaderConversionBenchmark  thrpt    5  13235.568 ± 279.687  ops/s
HttpResponseHeaderConversionBenchmark.http2HeaderConversionBenchmark  thrpt    5  14012.449 ± 208.878  ops/s
```

branch `ccbd589`
```
Benchmark                                                             Mode  Cnt      Score     Error  Units
HttpRequestHeaderConversionBenchmark.http1HeaderConversionBenchmark   thrpt    5  15048.927 ± 314.943  ops/s
HttpRequestHeaderConversionBenchmark.http2HeaderConversionBenchmark   thrpt    5  14633.153 ± 100.257  ops/s
HttpResponseHeaderConversionBenchmark.http1HeaderConversionBenchmark  thrpt    5  14790.639 ± 267.539  ops/s
HttpResponseHeaderConversionBenchmark.http2HeaderConversionBenchmark  thrpt    5  14854.848 ± 396.054  ops/s
```