@minwoox 2020-02-06
Are we going to separate the classes in this PR? 😄