@jyblue 2020-02-08
> Are we going to separate the classes in this PR?

It would be nice we handle this another PR to avoid more conflicts with master branch.
I will update PR this weekends :grinning: 
Thank you for your patience, reviewers :bowing_man: