@minwoox 2020-02-10
> It would be nice we handle this another PR to avoid more conflicts with master branch.

Really sorry about the conflicts. 😆
However, as I left the comment before, I'm afraid of having so many branches in the encoder classes and it's really hard to review.
Those classes are critical components in Armeria because every request and response has to go through them so I think we should fix it before we merge.
We cannot release a new version until we fix that if we merge it. 😄 

So could you refactor the classes in this PR, please?
Sorry for the conflicts again, and I will solve the conflicts if it happens again at this time. 😆