@minwoox 2020-02-10
>  also sorry for misunderstanding your comment.

Sorry for making you get confused. 😆 

> So I was worry about too much delays of this PR.

You don't have to worry about it. This will improve performance a lot in certain circumstances, but it's not a critical feature we need right now. So we can take our time. 😄 No hurry please. 😆