@jyblue 2020-02-12
This is in-progress commit. 
I will also split `ArmeriaHttpUtil` and do more clean up soon :grinning: