@jyblue 2020-02-26
Hello, here is update.
'Http1ObjectEncoder', 'Http2ObjectEncoder' , 'ArmeriaHttpUtil.toNettyHttp1()', 'ArmeriaHttpUtil.toNettyHttp2()' are now split.
Please take a look 😄
Thank you in advance.