@trustin 2020-03-17
This has been open for too long and getting more and more conflicts. We obviously do not want @jyblue to have poor experience contributing, so I'd like to merge this after fixing the merge conflict this time. `writeHeaders` could be split in a follow-up PR.